LipiPlayer/lipiplayer-webapp/requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.41.1
numpy==2.2.6
packaging==24.2
pandas==2.3.0
pillow==11.2.1
praat-parselmouth==0.4.5
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
pydub==0.25.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
reportlab==4.4.1
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
